Update vm_install_using_vagrant.md
Ubuntu is very behind on VirtualBox and Vagrant updates, so much that the current versions available on the main repositories don't work. I have updated the Ubuntu documentation to reflect this and add new repositories (Official VirtualBox Repo and Unofficial Vagrant Repo) in order to install the latest versions available. I would like to note that this installs Vagrant 2.0.3 and VirtualBox 5.2.8 (at this point in time) which is above the recommended versions. I have personally tested both of these versions (on Windows) and haven't encountered any issues with these versions.