Add installation instructions of ChromeDriver on Linux and Windows (#73)
* Include installation of ChromeDriver on Linux

Clear and concise documentation of the installation of ChromeDriver is not available on the website linked for ChromeDriver. I was using it for the first time and got fairly confused, having use an external blog to understand its installation process. It will help users to have it here. They keep changing the version numbers and there's no 'latest' release directory, thus the user may have to write the latest download's URL/version number himself.

* Update end_to_end_tests.md