[Documentation:Developer] set git credentials working with forks (#655)
**Summary:**
Added a small warning in the "Commit to PR from fork" page that tells
developers to ensure that their git credentials are set before pulling
and pushing from forks. If credentials are not set, then some commands
will fail. This change would help newer developers who might not be
proficient with git.

**Testing**
- Verified that changes appeared on the page
- Format looks good