Dark Mode Toggles
Dark mode now toggles with a 200ms transition time. TODO: Tidy up using CSS variables and not hard coding. Swap any images/icons or things that don't contrast well with dark mode. Style the button to be a nice clean slider. Make sure button placement is clean on mobile, unsure of how to implement that but will figure it out.