[Documentation:InstructorUI] Add redaction instructions (#682)
This pull request introduces a new feature for handling redactions in
bulk PDF uploads to enhance student privacy during grading. The changes
include documentation updates explaining how to prepare and view
redactions.

### Redaction Feature Documentation:

*
[`_docs/instructor/assignment_preparation/bulk_pdf_upload.md`](diffhunk://#diff-1c99d839d2dfcea1c6b0d4f30171cd40bc8e53df5768e0e27613c031e42c872aR256-R291):
Added a section on redactions, detailing the creation of a JSON file for
specifying redacted areas on PDF pages, enabling the feature in the
"Rubric" tab, and viewing redactions during grading. Includes examples
of JSON format and visual aids for preparation and viewing.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: Barb Cutler <Barb Cutler>